Rename retrieval dataset module to dataset_retrieval.
Avoids naming collision with backend.retrieval.
